Fock state basis set update
Jayson provided his MATLAB code that can be used for arbitrary number of sites. The code was bear of any documentation and so I added one as well. I also changed some notation in zeroeye and columnofones without affecting the results; the changes I made were to make this more intuitive when putting arguments in those functions.